Incorporate the ECOMP SDC Artefact Generator code
Remove the Maven dependency on openecomp-sdc-artifact-generator-core,
because this is not maintained in the ONAP source code repositories.
The functionality provided by the SDC Artifact Generator is replaced
with equivalent code that depends on the SDC-TOSCA parsing library,
without introducing any backwards-compatibility breaks.
The intention is to make this Maven project more maintainable by
depending only the common ONAP libraries.

Change-Id: I01d78575d3b7c70a11e4c7989a021de3c0913a06
Issue-ID: AAI-943
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>