Update patch set 1
Patch Set 1:

> Why has this change been made? Are there any security implications
 > for allowing the container to run as root and/or allowing users to
 > log into the image as root?
 > 
 > Is it an established ONAP convention for all images to run as root?

It is the convention in AAI to have the containers run as root (babel is the only one that doesn't).  If it doesn't run as root, it means that the deployment infrastructure needs to ensure that the user that it does run as (aaiadmin) is created on the host VM.  I don't believe OOM does this ... which will mean that babel won't have permissions to write to the log volume.

I'm open to a discussion about having all containers run as something other than root (and then changing OOM to ensure that this user exists).  But for now, aligning babel with the rest of the services to run as root seems to make sense.

Patch-set: 1
Reviewer: Gerrit User 453 <453@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0