Update SdWanServiceTest CSAR to make it valid
The SD-WAN test CSAR file is parseable using the sdc-tosca dependency
version 1.4.1 but it cannot be parsed using version 1.4.2 and above, as
there are missing data types leading to inheritance errors.

Correct the test CSAR file by adding the missing types. This has a
beneficial side-effect of increasing code coverage.

Issue-ID: AAI-1250
Change-Id: I41682b49d9d0a260a6feff05fda135949b46ec05
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>