Load type mappings from a group configuration file
Change the filter-types.properties file: re-implement to use JSON
content. Remove the System Property that defines the location of the
file and add this to the Spring application properties.

Initialise the type-to-model mappings using the key=value pairs from
this JSON content.

Refactor existing Junit tests to remove duplicated resource loading
code.

Change-Id: Idb2e962fe5cae39b70cc8cf16053d0a253f4fac0
Issue-ID: AAI-2121
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>