Convert filter-types.properties to an INI file
Use Commons Configuration2 to parse the existing properties file using
the INI parser. This allows the instance-group-types filter list to be specified on
separate lines (instead of comma-separated).

It also allows for additional INI sections to be specified (in a future
change), so that TOSCA to Java mappings can also be stored in this file.

Change-Id: I186ae7785443e2928c4394181a2b3fc58ac2268b
Issue-ID: AAI-2121
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>