Re-implement model type value for Resource Mapping
Refactor the Model/Service/Resource/Widget hierarchy to improve code
readability.
Replace the configurable Boolean value for a Resource's model type with
a value from the existing ModelType enum.

Obtain the A&AI XML Model name (always resource or service) directly
from the Model class and not from the configured model type for the
Resource.

Change-Id: I1487d919392b34e4bf1314da8cd2805a9fbf96a3
Issue-ID: AAI-2212
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>