Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Is there a reason the logback.xml was added to src/main/resources? Currently the logback.xml file is a part of the kubernetes config map, so that the logging levels can be configured externally. Will this change prevent that?
> 
> The main reason I've added it is I thought this file is missed. I see it exists in the other projects. Not sure which config will have higher precedence in case of configMap.

I think once Spring Boot has initialised, it will use the logback in the config folder, since it is configured to do so in application.properties:

    logging.config=${CONFIG_HOME}/logback.xml

However, any logging that happens before Spring Boot initialisation will pickup the logback from the classpath, so the one in src/main/resources will be used.

It might be worth making this clear in a comment for example in your logback.xml, that it is not the main one and is only used during Spring Boot initialisation; so that users don't naively update it expecting their changes to take effect.

Patch-set: 3
Label: Code-Review=0