Update patch set 4
Patch Set 4: Code-Review+1

reviewed, looks good. The setOrchestratrionType could get the "Model" between "set" and "orchestrationType", to be in line with other setters in the AaiModelGenerator.java, but this is not really required for this code to work properly.

Patch-set: 4
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1